+++ authentik/rbac/api/rbac_assigned_by_roles.py
@@ -1,5 +1,6 @@
 """common RBAC serializers"""
 
+from django.apps import apps
 from django.contrib.auth.models import Permission
 from django.db.models import Q, QuerySet
 from django.db.transaction import atomic
@@ -9,7 +10,8 @@ from drf_spectacular.utils import OpenApiResponse, extend_schema
 from guardian.models import RoleModelPermission, RoleObjectPermission
 from guardian.shortcuts import assign_perm, remove_perm
 from rest_framework.decorators import action
-from rest_framework.fields import CharField, ReadOnlyField
+from rest_framework.exceptions import ValidationError
+from rest_framework.fields import CharField, ChoiceField, ListField, ReadOnlyField
 from rest_framework.mixins import ListModelMixin
 from rest_framework.request import Request
 from rest_framework.response import Response
@@ -17,9 +19,48 @@ from rest_framework.viewsets import GenericViewSet
 
 from authentik.core.api.utils import ModelSerializer, PassiveSerializer
 from authentik.lib.api import model_choices
-from authentik.rbac.api.rbac import PermissionAssignResultSerializer, PermissionAssignSerializer
+from authentik.lib.validators import RequiredTogetherValidator
 from authentik.rbac.decorators import permission_required
 from authentik.rbac.models import Role
+
+
+class PermissionAssignResultSerializer(PassiveSerializer):
+    """Result from assigning permissions to a user/role"""
+
+    id = CharField()
+
+
+class PermissionAssignSerializer(PassiveSerializer):
+    """Request to assign a new permission"""
+
+    permissions = ListField(child=CharField())
+    model = ChoiceField(choices=model_choices(), required=False)
+    object_pk = CharField(required=False)
+
+    validators = [RequiredTogetherValidator(fields=["model", "object_pk"])]
+
+    def validate(self, attrs: dict) -> dict:
+        model_instance = None
+        # Check if we're setting an object-level perm or global
+        model = attrs.get("model")
+        object_pk = attrs.get("object_pk")
+        if model and object_pk:
+            model = apps.get_model(attrs["model"])
+            model_instance = model.objects.filter(pk=attrs["object_pk"]).first()
+        attrs["model_instance"] = model_instance
+        if attrs.get("model"):
+            return attrs
+        permissions = attrs.get("permissions", [])
+        if not all("." in perm for perm in permissions):
+            raise ValidationError(
+                {
+                    "permissions": (
+                        "When assigning global permissions, codename must be given as "
+                        "app_label.codename"
+                    )
+                }
+            )
+        return attrs
 
 
 class RoleObjectPermissionSerializer(ModelSerializer):
